Quitting in davinci with modules loaded was causing SEGV due to double free during top-level scope cleanup. Fixed by unloading modules before cleaning the Symtable.
The double-free happened when the top-level Symtable was cleaned (via clean_table)
during scope cleanup. The cleanup triggered module unload due to free_var(). 
Module unload automatically cleared the symbol for the module and freed it.
The clean_table routine not knowing that the Symtable had been alreaded freed,
freed it yet again.



git-svn-id: http://oss.mars.asu.edu/svn/davinci/davinci/trunk@18629 dedd6328-c50f-0410-8be5-b54a49638459