Addressed the following bugs related to user defined functions: 1. It is possible to overwrite actual parameters passed to user    defined functions within the body of the user defined function.    (Bug 1919). 2. Actual parameters passed as values of named formal parameters    were being resolved in the user function scope in certain cases.    (Bug 1951). For example, "foo(x=d)" call would try to resolve "d"    within foo's body.
The bug was fixed by duplicating the actual parameter "a" passed
to the user defined function during dispatch (see ufunc.c) insted
of just copying the reference. The duplication removed the possibility
of corrupting the (actual) input variable.

In addition, calling "foo(x=a)", which translated to eval("x")
C-function, which cascaded to eval("a"), but in the wrong scope
(i.e. function's scope). That resulted in the error message: "error:
Variable not found: a" during function execution.



git-svn-id: http://oss.mars.asu.edu/svn/davinci/davinci/trunk@19141 dedd6328-c50f-0410-8be5-b54a49638459