Bug 2169, load_csv() does not promote floating point data to doubles
This is in the file dvio_csv.c

Saadat suggested that floats be retained where feasible.
I added a test to see if the float and the double conversions differed by more than 0.5.

I picked 0.5 because it is a good idea to keep sclk times within 1 second.

Thus, if the difference in a conversion is > 0.5, the double is used.  If it is less, the float is used

git-svn-id: http://oss.mars.asu.edu/svn/davinci/davinci/trunk@19413 dedd6328-c50f-0410-8be5-b54a49638459