Fix HDF dimension ordering issue and a small memory leak
Since dvio_hdf.c was added to davinci we've been reading the dimensions backwards.
This apparently hasn't bothered anyone even though our own hdf files (in the repo)
whith our extra "org" attribute stating the file was BSQ (XYZ) would actually be
ZYX.  There's a comment in the code with more detail/explanation.

I also got rid of the HDF5 error output for missing "org" and "lines" attributes
by checking for existence first (so only our "parse_error" is still there if they're
missing) and I also changed it so we write data out in native format not always BE
because we can read in the data either way so no reason to incur the extra overhead
on LE machines which is 99.99999% of use.

Also fixed a minor memory leak in ff_write.c 


git-svn-id: http://oss.mars.asu.edu/svn/davinci/davinci/trunk@19508 dedd6328-c50f-0410-8be5-b54a49638459