Add HDF5 reader/writer logic and tests
Now we have several paths:

if hdf_old is set for read or write, we'll do it exactly like we always
have in the past.

otherwise, if the "org" attribute is found and "dv_std" is not, we'll read it like
we always did but we'll write it out fixed, with the new "dv_std" attribute.

If dv_std is found we read it in the new, correct way.

If "org" isn't found we read it the new correct way since it obviously wasn't
writted by davinci.

The upshot of this mess is backwards compatibility and no changes necessary for almost
every use case or at most merely adding hdf_old=1 to the calls.

I've added a couple of hdf tests and moved them and associated test files to their own folder.

I also fixed Don's Advanced Fits test so it doesn't reference his home directory and it actually
returns a failure if it fails.



git-svn-id: http://oss.mars.asu.edu/svn/davinci/davinci/trunk@19513 dedd6328-c50f-0410-8be5-b54a49638459